+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Introduce empty compute plan status (#726)
 
+### Changed
+- base docker image from alpine 3.15 to alpine 3.16 (#751)
+
 ### Fixed
 - event asset migration (#750)